+++ example/bin/example.dart
@@ -1,120 +1,88 @@
 import 'package:geojson_vi/geojson_vi.dart';
 
 Future<void> main() async {
+  // Create GeoJSONPoint
+  final geoJSONPoint = GeoJSONPoint([105.77389, 21.0720414]);
+  print('GeoJSONPoint:\n${geoJSONPoint.toMap()}');
+  print('Area: ${geoJSONPoint.area}');
+  print('Distance: ${geoJSONPoint.distance}');
+
+  // Create GeoJSONPoint from Map
   final pointDataMap = {
     'type': 'Point',
     'coordinates': [105.77389, 21.0720414]
   };
+  final geoJSONPointFromMap = GeoJSONPoint.fromMap(pointDataMap);
+  print('\nGeoJSONPoint from Map:\n${geoJSONPointFromMap.toMap()}');
 
-  /// Create GeoJSON object from data without knowing its type
-  final geoJSONPoint = GeoJSON.fromMap(pointDataMap);
-  if (geoJSONPoint is GeoJSONPoint) {
-    final type = geoJSONPoint.type;
-    final area = geoJSONPoint.area;
-    final bbox = geoJSONPoint.bbox;
-    final coordinates = geoJSONPoint.coordinates;
-    final distance = geoJSONPoint.distance;
-    final map = geoJSONPoint.toMap();
-    final json = geoJSONPoint.toJSON();
-    final string = geoJSONPoint.toString();
-    print('$type\n$area\n$bbox\n$coordinates\n'
-        '$distance\n$map\n$json\n$string');
+  // Create GeoJSONPoint from JSON
+  final pointJSON = '{"type": "Point", "coordinates": [105.77389, 21.0720414]}';
+  final geoJSONPointFromJSON = GeoJSONPoint.fromJSON(pointJSON);
+  print('\nGeoJSONPoint from JSON:\n${geoJSONPointFromJSON.toMap()}');
 
-    /// Create GeoJSONPoint object from JSON string already know its type
-    final geoJSONPoint1 = GeoJSONPoint.fromJSON(json);
-    final type1 = geoJSONPoint1.type;
-    final area1 = geoJSONPoint1.area;
-    final bbox1 = geoJSONPoint1.bbox;
-    final coordinates1 = geoJSONPoint1.coordinates;
-    final distance1 = geoJSONPoint1.distance;
-    final map1 = geoJSONPoint1.toMap();
-    final json1 = geoJSONPoint1.toJSON();
-    final string1 = geoJSONPoint1.toString();
-    print('$type1\n$area1\n$bbox1\n$coordinates1\n'
-        '$distance1\n$map1\n$json1\n$string1');
-  }
-
-  /// Create GeoJSONGeometry object from data without knowing its type
-  final geoJSONGeometry = GeoJSONGeometry.fromMap(pointDataMap);
-  if (geoJSONGeometry is GeoJSONPoint) {
-    final type = geoJSONGeometry.type;
-    final area = geoJSONGeometry.area;
-    final bbox = geoJSONGeometry.bbox;
-    final coordinates = geoJSONGeometry.coordinates;
-    final distance = geoJSONGeometry.distance;
-    final map = geoJSONGeometry.toMap();
-    final json = geoJSONGeometry.toJSON();
-    final string = geoJSONGeometry.toString();
-    print('$type\n$area\n$bbox\n$coordinates\n'
-        '$distance\n$map\n$json\n$string');
-
-    /// Create GeoJSONPoint object from JSON string already know its type
-    final geoJSONGeometr1 = GeoJSONPoint.fromJSON(json);
-    final type1 = geoJSONGeometr1.type;
-    final area1 = geoJSONGeometr1.area;
-    final bbox1 = geoJSONGeometr1.bbox;
-    final coordinates1 = geoJSONGeometr1.coordinates;
-    final distance1 = geoJSONGeometr1.distance;
-    final map1 = geoJSONGeometr1.toMap();
-    final json1 = geoJSONGeometr1.toJSON();
-    final string1 = geoJSONGeometr1.toString();
-    print('$type1\n$area1\n$bbox1\n$coordinates1\n'
-        '$distance1\n$map1\n$json1\n$string1');
-  }
-
-  /// Creating a new feature collection
-  final newFeatureColl = GeoJSONFeatureCollection([]);
-
-  final point = GeoJSONPoint([105.780701, 21.067921]);
-  final pointFeature = GeoJSONFeature(point);
-  pointFeature.properties = {
-    'title': 'Hanoi University of Mining and Geology',
-    'url': 'http://humg.edu.vn',
-  };
-
-  final polygonCoordinates = [
+  // Create GeoJSONPolygon
+  final geoJSONPolygon = GeoJSONPolygon([
     [
-      [104.765625, 20.468189],
-      [106.545410, 20.468189],
-      [106.545410, 21.596150],
-      [104.765625, 21.596150],
-      [104.765625, 20.468189]
+      [105.77389, 21.0720414],
+      [105.77123, 21.074412],
+      [105.77015, 21.075051],
+      [105.77389, 21.0720414]
     ]
-  ];
-  newFeatureColl.features.add(pointFeature);
-  final polygon = GeoJSONPolygon(polygonCoordinates);
-  final polygonFeature = GeoJSONFeature(polygon);
-  newFeatureColl.features.add(polygonFeature);
-  print(newFeatureColl.toMap());
-  /*
-    {
-    'type': 'FeatureCollection',
-    'features': [
-      {
-        'type': 'Feature',
-        'properties': {'name': 'This is a Point'},
-        'geometry': {
-          'type': 'Point',
-          'coordinates': [105.780163, 21.067921]
-        }
-      },
-      {
-        'type': 'Feature',
-        'properties': {'name': 'Example of Polygon'},
-        'geometry': {
-          'type': 'Polygon',
-          'coordinates': [
-            [
-              [104.765625, 20.468189],
-              [106.54541, 20.468189],
-              [106.54541, 21.59615],
-              [104.765625, 21.59615],
-              [104.765625, 20.468189]
-            ]
-          ]
-        }
-      }
+  ]);
+  print('\nGeoJSONPolygon:\n${geoJSONPolygon.toMap()}');
+  print('Area: ${geoJSONPolygon.area}');
+  print('Distance: ${geoJSONPolygon.distance}');
+
+  // Create GeoJSONPolygon from Map
+  final polygonDataMap = {
+    'type': 'Polygon',
+    'coordinates': [
+      [
+        [105.77389, 21.0720414],
+        [105.77123, 21.074412],
+        [105.77015, 21.075051],
+        [105.77389, 21.0720414]
+      ]
     ]
+  };
+  final geoJSONPolygonFromMap = GeoJSONPolygon.fromMap(polygonDataMap);
+  print('\nGeoJSONPolygon from Map:\n${geoJSONPolygonFromMap.toMap()}');
+
+  // Create GeoJSONPolygon from JSON
+  final polygonJSON = '''
+  {
+    "type": "Polygon",
+    "coordinates": [[
+      [105.77389, 21.0720414],
+      [105.77123, 21.074412],
+      [105.77015, 21.075051],
+      [105.77389, 21.0720414]
+    ]]
   }
-  */
+  ''';
+  final geoJSONPolygonFromJSON = GeoJSONPolygon.fromJSON(polygonJSON);
+  print('\nGeoJSONPolygon from JSON:\n${geoJSONPolygonFromJSON.toMap()}');
+
+  // Create GeoJSONFeatureCollection
+  final geoJSONFeatureCollection = GeoJSONFeatureCollection([
+    GeoJSONFeature(geoJSONPoint, properties: {'name': 'My Point'}),
+    GeoJSONFeature(geoJSONPolygon, properties: {'name': 'My Polygon'}),
+  ]);
+  print('\nGeoJSONFeatureCollection:\n${geoJSONFeatureCollection.toMap()}');
+
+  // Find features
+  print('\nFind features with name "My Point":');
+  final foundFeaturesExact = geoJSONFeatureCollection
+      .findProperties('name', 'My Point', contains: false);
+  for (var feature in foundFeaturesExact) {
+    print(feature.toMap());
+  }
+
+  // Find features using containment
+  print('\nFind features with name containing "My":');
+  final foundFeaturesContainment =
+      geoJSONFeatureCollection.findProperties('name', 'My', contains: true);
+  for (var feature in foundFeaturesContainment) {
+    print(feature.toMap());
+  }
 }